minor bug

+++ ChamberAcceptance.cc
@@ -28,8 +28,8 @@ void GetAcceptance(TString var_name, TString kinematics, TString hist_name, TStr
     auto df_mix = df_mix_with_basic_cuts.Filter(mix_cuts.Data());
     auto df_mc = df_mc_with_basic_cuts.Filter(mc_cuts.Data());
 
-    auto xmin = df_real.Min<double>(var_name.Data());
-    auto xmax = df_real.Max<double>(var_name.Data());
+    auto xmin = df_real.Min<float>(var_name.Data());
+    auto xmax = df_real.Max<float>(var_name.Data());
 
 
     TString hReal_name = Form("hist_%s_real", hist_name.Data());
